requirements.txt
fastapi
uvicorn[standard]
daytona-sdk
pyyaml
